[Maplab] Refactor.
