+++ src/openschema_alchemy/maplab/protobuf/aslam_serialization/camera.proto
