+++ src/openschema_alchemy/maplab/protobuf/aslam_serialization/visual_frame.proto
