+++ src/openschema_alchemy/maplab/protobuf/summary_map.proto
