+++ src/openschema_alchemy/maplab/protobuf/vi_map/sensor_resources.proto
